hw 4.5 дз готово

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -97,7 +97,7 @@ public class StudentController {
     }
 
     @GetMapping("a+b")
-    public int iint(){
+    public long iint(){
         return studentService.iint();
     }
